#3615 comment: code written; tests pending
