+++ src/main.py
@@ -32,173 +32,7 @@ import thiscovery_lib.utilities as utils
 from thiscovery_lib.dynamodb_utilities import Dynamodb
 from common.constants import STACK_NAME, STATUS_TABLE, AUDIT_TABLE, PROJECTS_TABLE
 from common.mediaconvert_utilities import MediaConvertClient
-
-
-class IncomingMonitor:
-
-    def __init__(self, logger, correlation_id=None):
-        self.logger = logger
-        self.correlation_id = correlation_id
-        self.ddb_client = Dynamodb(stack_name=STACK_NAME)
-        self.s3_client = S3Client()
-        self.known_files = None
-        self.active_projects = None
-        self.core_api_url = None
-
-    def get_core_api_url(self):
-        env_name = utils.get_environment_name()
-        if env_name == 'prod':
-            return 'https://api.thiscovery.org/'
-        else:
-            return f'https://{env_name}-api.thiscovery.org/'
-
-    def get_user_id_from_core_api(self, email):
-        result = utils.aws_get('v1/user', self.get_core_api_url(), params={'email': email})
-        assert result['statusCode'] == HTTPStatus.OK, f'Call to core API returned error: {result}'
-        return json.loads(result['body'])['id']
-
-    def get_active_projects_info(self):
-        self.active_projects = self.ddb_client.scan(
-            table_name=PROJECTS_TABLE,
-            filter_attr_name="interview_task_status",
-            filter_attr_values=['active']
-        )
-        return self.active_projects
-
-    def get_anon_project_specific_user_id(self, user_id, project_id):
-        result = utils.aws_get('v1/userproject', self.get_core_api_url(), params={
-            'user_id': user_id
-        })
-        assert result['statusCode'] == HTTPStatus.OK, f'Call to core API returned error: {result}'
-        user_projects = json.loads(result['body'])
-        for up in user_projects:
-            if up['project_id'] == project_id:
-                return up['anon_project_specific_user_id']
-        raise utils.ObjectDoesNotExistError(f'User project not found', details={
-            'user_id': user_id,
-            'project_id': project_id,
-            'correlation_id': self.correlation_id,
-        })
-
-    def add_new_file_to_status_table(self, s3_bucket_name, s3_path, head):
-        if self.active_projects is None:
-            self.get_active_projects_info()
-        if not self.active_projects:
-            raise utils.ObjectDoesNotExistError(f"No research projects conducting interviews could be found in Dynamodb {PROJECTS_TABLE} table",
-                                                details={'active_projects': self.active_projects})
-
-        self.logger.debug('Path of S3 file', extra={'s3_path': s3_path, 's3_bucket_name': s3_bucket_name})
-        s3_filename, interview_dir, file_type = parse_s3_path(s3_path)
-        metadata = head['Metadata']
-        user_id = self.get_user_id_from_core_api(metadata['email'])
-        referrer_url = metadata.get('referrer')
-        project_acronym = None
-        project_prefix = None
-        project_id = None
-
-        if referrer_url:  # on-demand interview
-            interview_type = 'INT-O'
-            question_number = metadata['question_index']
-            for p in self.active_projects:
-                if p.get('on_demand_referrer') == referrer_url:
-                    project_acronym = p['id']
-                    project_prefix = p['filename_prefix']
-                    project_id = p['project_id']
-                    break
-            assert project_prefix, f'Referrer url {referrer_url} not found in Dynamodb table {PROJECTS_TABLE}'
-            anon_project_specific_user_id = self.get_anon_project_specific_user_id(
-                user_id=user_id,
-                project_id=project_id,
-            )
-            target_basename = f'{project_prefix}_{interview_type}_{anon_project_specific_user_id}_{question_number}'
-        else:  # live interview
-            interview_type = 'INT-L'
-            interviewer = metadata['interviewer']
-            if len(self.active_projects) == 1:
-                project = self.active_projects[0]
-            else:
-                interviewer_matches = list()
-                for p in self.active_projects:
-                    if interviewer in p['interviewers'].keys():
-                        interviewer_matches.append(p)
-                assert len(interviewer_matches) == 1, f'Could not resolve project based on interviewer. Interviewer {interviewer} is taking part in ' \
-                                                      f'projects: {", ".join(map(lambda m: m["id"], interviewer_matches))}'
-                project = interviewer_matches[0]
-            project_acronym = project['id']
-            project_prefix = project['filename_prefix']
-            project_id = project['project_id']
-            interviewer_initials = project['interviewers'][interviewer]['initials']
-            anon_project_specific_user_id = self.get_anon_project_specific_user_id(
-                user_id=user_id,
-                project_id=project_id,
-            )
-            target_basename = f'{project_prefix}_{interview_type}_{interviewer_initials}_{anon_project_specific_user_id}'
-
-        item = {
-            'original_filename': s3_filename,
-            'original_path': s3_path,
-            'source_bucket': s3_bucket_name,
-            'interview_id': interview_dir,
-            'project_acronym': project_acronym,
-            'target_basename': target_basename,
-            'audio_extraction_attempts': 0,
-            'sdhs_transfer_attempts': 0,
-            'processing_status': 'new',
-        }
-        head['uploaded_to_s3'] = str(head.get('LastModified'))
-        del head['LastModified']
-        try:
-            self.logger.debug('Adding item to FileTransferStatus table', extra={'item': item})
-            result = self.ddb_client.put_item(
-                table_name=STATUS_TABLE,
-                key=s3_path,
-                item_type=file_type,
-                item_details=head,
-                item=item,
-                correlation_id=self.correlation_id
-            )
-            assert result['ResponseMetadata']['HTTPStatusCode'] == HTTPStatus.OK, f"put_item operation failed with response: {result}"
-            return result
-        except utils.DetailedValueError:
-            self.logger.error(f'Key {s3_path} already exists in DynamoDb table', extra={'key': s3_path})
-            raise
-
-    def main(self, ignore_extensions=['.mp3', '.flac'], bucket_name=None):
-        """
-        The main processing routine
-
-        Args:
-            ignore_extensions (list): list of file extensions to ignore
-            bucket_name (str): specify a different target bucket; used for testing
-
-        Returns:
-        """
-        files_added_to_status_table = list()
-        self.known_files = [x['id'] for x in self.ddb_client.scan(STATUS_TABLE)]
-        if bucket_name:
-            s3_bucket_name = f'{STACK_NAME}-{utils.get_environment_name()}-{bucket_name}'
-        else:
-            s3_bucket_name = utils.get_secret("incoming-interviews-bucket", namespace_override='/prod/')['name']
-            # # the approach in the line below does not work on AWS; use a S3 bucket policy to allow access from this lambda instead.
-            # self.s3_client = S3Client(utils.namespace2profile('/prod/'))  # use an s3_client with access to production buckets
-        objs = self.s3_client.list_objects(s3_bucket_name)['Contents']
-        for o in objs:
-            s3_path = o['Key']
-            folder = s3_path.split('/')[0]
-
-            # skip anything that is not in a uuid-named folder
-            try:
-                utils.validate_uuid(folder)
-            except utils.DetailedValueError:
-                continue
-
-            if s3_path not in self.known_files:
-                _, extension = os.path.splitext(s3_path)
-                if extension and (extension not in ignore_extensions):
-                    head = self.s3_client.head_object(s3_bucket_name, s3_path)
-                    self.add_new_file_to_status_table(s3_bucket_name, s3_path, head)
-                    files_added_to_status_table.append(s3_path)
-        return files_added_to_status_table
+from common.helpers import parse_s3_path
 
 
 class ProcessIncoming:
@@ -373,12 +207,6 @@ class Cleaner:
         self.clean_incoming_bucket()
         self.clean_audio_bucket()
         self.move_items_to_audit_table()
-
-
-def parse_s3_path(s3_path):
-    s3_dirs, s3_filename = os.path.split(s3_path)
-    interview_dir, file_type = s3_dirs.split('/')[-2:]
-    return s3_filename, interview_dir, file_type
 
 
 @utils.lambda_wrapper
